pull some charts from jupyter notebook into a standalone script.  Make jupyter deps installation easier.  Update jupyter notebook with new charts.  Remove Pipfile.
