--- Pipfile
@@ -1,16 +1,0 @@
-[[source]]
-name = "pypi"
-url = "https://pypi.org/simple"
-verify_ssl = true
-
-[dev-packages]
-
-[packages]
-matplotlib = "3.2.1"
-pandas = "1.0.3"
-numpy = "1.18.2"
-durationpy = "0.5"
-adjusttext = "0.7.3"
-
-[requires]
-python_version = "3.6"
