+++ benchmark_results_data/evm_benchmarks.csv
@@ -1,0 +1,101 @@
+engine,test_name,total_time,gas_used
+evmone,sha1-10808-bits-shiftopt,0.001596,759.785k
+parity-evm,sha1-10808-bits-shiftopt,0.02231382,759785
+geth-evm,sha1-10808-bits-shiftopt,0.017769533,759785
+cita-evm,sha1-10808-bits-shiftopt,0.012675902,873117
+evmone,sha1-21896-bits-shiftopt,0.003226,1.48425M
+parity-evm,sha1-21896-bits-shiftopt,0.04014344,1484254
+geth-evm,sha1-21896-bits-shiftopt,0.035940819,1484254
+cita-evm,sha1-21896-bits-shiftopt,0.025491886,1691062
+evmone,sha1-42488-bits-shiftopt,0.006501,2.89847M
+parity-evm,sha1-42488-bits-shiftopt,0.086441834,2898473
+geth-evm,sha1-42488-bits-shiftopt,0.076433283,2898473
+cita-evm,sha1-42488-bits-shiftopt,0.050729618,3280257
+evmone,blake2b-8415-bytes-shiftopt,0.037586,12.0431M
+parity-evm,blake2b-8415-bytes-shiftopt,0.304234694,12043105
+geth-evm,blake2b-8415-bytes-shiftopt,0.250975,12043105
+cita-evm,blake2b-8415-bytes-shiftopt,0.227318166,12635513
+evmone,blake2b-5610-bytes-shiftopt,0.024851,7.43482M
+parity-evm,blake2b-5610-bytes-shiftopt,0.191787707,7434820
+geth-evm,blake2b-5610-bytes-shiftopt,0.164829978,7434820
+cita-evm,blake2b-5610-bytes-shiftopt,0.148494879,7837276
+evmone,blake2b-2805-bytes-shiftopt,0.012214,3.42353M
+parity-evm,blake2b-2805-bytes-shiftopt,0.099552405,3423526
+geth-evm,blake2b-2805-bytes-shiftopt,0.08373404,3423526
+cita-evm,blake2b-2805-bytes-shiftopt,0.075647999,3635902
+evmone,blake2b-8415-bytes,0.044307,12.3518M
+parity-evm,blake2b-8415-bytes,0.312272394,12351815
+geth-evm,blake2b-8415-bytes,0.258288143,12351815
+cita-evm,blake2b-8415-bytes,0.287592494,12944223
+evmone,blake2b-5610-bytes,0.029281,7.64067M
+parity-evm,blake2b-5610-bytes,0.197467613,7640669
+geth-evm,blake2b-5610-bytes,0.176870627,7640669
+cita-evm,blake2b-5610-bytes,0.191670564,8043125
+evmone,blake2b-2805-bytes,0.015022,3.52651M
+parity-evm,blake2b-2805-bytes,0.1041588,3526514
+geth-evm,blake2b-2805-bytes,0.08633356,3526514
+cita-evm,blake2b-2805-bytes,0.094874499,3738890
+evmone,bn256g2mul-0xAshish,0.00226,455.757k
+parity-evm,bn256g2mul-0xAshish,0.019948515,455757
+geth-evm,bn256g2mul-0xAshish,0.015922419,455757
+cita-evm,bn256g2mul-0xAshish,0.042454018,486053
+evmone,bn128_mul_weierstrudel-cdetrio2,0.000521,47.165k
+parity-evm,bn128_mul_weierstrudel-cdetrio2,0.004141905,47165
+geth-evm,bn128_mul_weierstrudel-cdetrio2,0.002277002,47165
+cita-evm,bn128_mul_weierstrudel-cdetrio2,0.017047386,74437
+evmone,bn128_mul_weierstrudel-chfast1,0.000294,23.449k
+parity-evm,bn128_mul_weierstrudel-chfast1,0.001957099,23449
+geth-evm,bn128_mul_weierstrudel-chfast1,0.001065043,23449
+cita-evm,bn128_mul_weierstrudel-chfast1,0.007357496,49441
+evmone,bn128_mul_weierstrudel-chfast2,0.000458,39.337k
+parity-evm,bn128_mul_weierstrudel-chfast2,0.003425564,39337
+geth-evm,bn128_mul_weierstrudel-chfast2,0.001819083,39337
+cita-evm,bn128_mul_weierstrudel-chfast2,0.014044644,66865
+evmone,bn128_mul_weierstrudel-chfast3,0.000525,47.291k
+parity-evm,bn128_mul_weierstrudel-chfast3,0.004222961,47291
+geth-evm,bn128_mul_weierstrudel-chfast3,0.002231105,47291
+cita-evm,bn128_mul_weierstrudel-chfast3,0.017467266,74819
+evmone,bn128_mul_weierstrudel-cdetrio6,0.000541,47.561k
+parity-evm,bn128_mul_weierstrudel-cdetrio6,0.004162919,47561
+geth-evm,bn128_mul_weierstrudel-cdetrio6,0.002232338,47561
+cita-evm,bn128_mul_weierstrudel-cdetrio6,0.017366266,75089
+evmone,bn128_mul_weierstrudel-cdetrio7,0.000523,47.165k
+parity-evm,bn128_mul_weierstrudel-cdetrio7,0.00429027,47165
+geth-evm,bn128_mul_weierstrudel-cdetrio7,0.002412266,47165
+cita-evm,bn128_mul_weierstrudel-cdetrio7,0.018339345,74501
+evmone,bn128_mul_weierstrudel-cdetrio11,0.000513,47.561k
+parity-evm,bn128_mul_weierstrudel-cdetrio11,0.004136138,47561
+geth-evm,bn128_mul_weierstrudel-cdetrio11,0.00215742,47561
+cita-evm,bn128_mul_weierstrudel-cdetrio11,0.016865172,75089
+evmone,bn128_mul_weierstrudel-cdetrio12,0.000571,47.165k
+parity-evm,bn128_mul_weierstrudel-cdetrio12,0.004018633,47165
+geth-evm,bn128_mul_weierstrudel-cdetrio12,0.002310364,47165
+cita-evm,bn128_mul_weierstrudel-cdetrio12,0.018379059,74501
+evmone,blake2b_huff-8415-bytes,0.002448,879.92k
+parity-evm,blake2b_huff-8415-bytes,0.027908764,879920
+geth-evm,blake2b_huff-8415-bytes,0.035075896,879920
+cita-evm,blake2b_huff-8415-bytes,0.015139534,1471220
+evmone,blake2b_huff-5610-bytes,0.001544,586.839k
+parity-evm,blake2b_huff-5610-bytes,0.017326637,586839
+geth-evm,blake2b_huff-5610-bytes,0.023309606,586839
+cita-evm,blake2b_huff-5610-bytes,0.010120834,988103
+evmone,blake2b_huff-2805-bytes,0.000844,293.755k
+parity-evm,blake2b_huff-2805-bytes,0.008337344,293755
+geth-evm,blake2b_huff-2805-bytes,0.011073158,293755
+cita-evm,blake2b_huff-2805-bytes,0.005137996,504983
+evmone,sha1-10808-bits,0.003175,790.156k
+parity-evm,sha1-10808-bits,0.055919944,790156
+geth-evm,sha1-10808-bits,0.020203397,790156
+cita-evm,sha1-10808-bits,0.053636084,903488
+evmone,sha1-21896-bits,0.005928,1.54361M
+parity-evm,sha1-21896-bits,0.120435923,1543605
+geth-evm,sha1-21896-bits,0.0451866,1543605
+cita-evm,sha1-21896-bits,0.10864109,1750413
+evmone,sha1-42488-bits,0.01191,3.0144M
+parity-evm,sha1-42488-bits,0.223914963,3014404
+geth-evm,sha1-42488-bits,0.082095587,3014404
+cita-evm,sha1-42488-bits,0.203935243,3396188
+evmone,mul256-gcolvin-drag-race,0.011393,5.61028M
+parity-evm,mul256-gcolvin-drag-race,0.114632564,5610277
+geth-evm,mul256-gcolvin-drag-race,0.167468693,5610277
+cita-evm,mul256-gcolvin-drag-race,0.043008777,5635901
