+++ benchmark_results_data/geth_precompile_benchmarks.csv
@@ -1,0 +1,64 @@
+test_name,gas,time
+PrecompiledEcrecover/,3000,0.00021357
+PrecompiledSha256/128,108,8.14e-07
+PrecompiledRipeMD/128,1080,1.943e-06
+PrecompiledIdentity/128,27,2.3e-08
+PrecompiledModExp/eip_example1,13056,4.6175e-05
+PrecompiledModExp/eip_example2,13056,9.251e-06
+PrecompiledModExp/nagydani-1-square,204,2.519e-06
+PrecompiledModExp/nagydani-1-qube,204,3.426e-06
+PrecompiledModExp/nagydani-1-pow0x10001,3276,1.8762e-05
+PrecompiledModExp/nagydani-2-square,665,4.164e-06
+PrecompiledModExp/nagydani-2-qube,665,6.306e-06
+PrecompiledModExp/nagydani-2-pow0x10001,10649,3.8218e-05
+PrecompiledModExp/nagydani-3-square,1894,7.628e-06
+PrecompiledModExp/nagydani-3-qube,1894,1.3436e-05
+PrecompiledModExp/nagydani-3-pow0x10001,30310,8.8893e-05
+PrecompiledModExp/nagydani-4-square,5580,1.7321e-05
+PrecompiledModExp/nagydani-4-qube,5580,3.7656e-05
+PrecompiledModExp/nagydani-4-pow0x10001,89292,0.000240053
+PrecompiledModExp/nagydani-5-square,17868,4.5674e-05
+PrecompiledModExp/nagydani-5-qube,17868,0.00010544
+PrecompiledModExp/nagydani-5-pow0x10001,285900,0.000707403
+PrecompiledBn256Add/chfast1,150,1.7695e-05
+PrecompiledBn256Add/chfast2,150,1.7853e-05
+PrecompiledBn256Add/cdetrio1,150,1.296e-06
+PrecompiledBn256Add/cdetrio2,150,1.336e-06
+PrecompiledBn256Add/cdetrio3,150,1.332e-06
+PrecompiledBn256Add/cdetrio4,150,1.374e-06
+PrecompiledBn256Add/cdetrio5,150,1.323e-06
+PrecompiledBn256Add/cdetrio6,150,1.686e-06
+PrecompiledBn256Add/cdetrio7,150,1.692e-06
+PrecompiledBn256Add/cdetrio8,150,1.723e-06
+PrecompiledBn256Add/cdetrio9,150,1.682e-06
+PrecompiledBn256Add/cdetrio10,150,1.69e-06
+PrecompiledBn256Add/cdetrio11,150,1.7936e-05
+PrecompiledBn256Add/cdetrio12,150,1.799e-05
+PrecompiledBn256Add/cdetrio13,150,1.7819e-05
+PrecompiledBn256Add/cdetrio14,150,2.602e-06
+PrecompiledBn256ScalarMul/chfast1,6000,0.000122872
+PrecompiledBn256ScalarMul/chfast2,6000,0.000133921
+PrecompiledBn256ScalarMul/chfast3,6000,0.000128134
+PrecompiledBn256ScalarMul/cdetrio1,6000,0.000138337
+PrecompiledBn256ScalarMul/cdetrio6,6000,0.000137062
+PrecompiledBn256ScalarMul/cdetrio11,6000,0.000137917
+PrecompiledBn256Pairing/jeff1,113000,0.004242082
+PrecompiledBn256Pairing/jeff2,113000,0.004200379
+PrecompiledBn256Pairing/jeff3,113000,0.004194623
+PrecompiledBn256Pairing/jeff4,147000,0.005602025
+PrecompiledBn256Pairing/jeff5,147000,0.005692373
+PrecompiledBn256Pairing/jeff6,113000,0.004273775
+PrecompiledBn256Pairing/empty_data,45000,0.001347387
+PrecompiledBn256Pairing/one_point,79000,0.002785696
+PrecompiledBn256Pairing/two_point_match_2,113000,0.004257572
+PrecompiledBn256Pairing/two_point_match_3,113000,0.00420171
+PrecompiledBn256Pairing/two_point_match_4,113000,0.004207033
+PrecompiledBn256Pairing/ten_point_match_1,385000,0.015446948
+PrecompiledBn256Pairing/ten_point_match_2,385000,0.015540603
+PrecompiledBn256Pairing/ten_point_match_3,113000,0.004229468
+PrecompiledBn256Pairing/two_point_match_rollup_input,113000,0.004204822
+PrecompiledBlake2F/vector_4,0,1.49e-07
+PrecompiledBlake2F/vector_5,12,3.06e-07
+PrecompiledBlake2F/vector_6,12,3.13e-07
+PrecompiledBlake2F/vector_7,1,1.63e-07
+PrecompiledBlake2F/vector_8,8000000,0.101986515
