+++ evm/scripts/run_bench.sh
@@ -13,5 +13,5 @@ docker run --env PYTHONIOENCODING=UTF-8 -v $(pwd)/evmraceresults:/evmraceresults
 docker run --env PYTHONIOENCODING=UTF-8 -v $(pwd)/evmraceresults:/evmraceresults -v $(pwd)/scripts:/scripts -v $(pwd)/input_data:/input_data -it cita-vm-bench /usr/bin/python3 /scripts/benchevm.py cita-vm
 
 # Merge benchmarks
-docker run --env PYTHONIOENCODING=UTF-8 -v $(pwd)/evmraceresults:/evmraceresults -v $(pwd)/scripts:/scripts -v $(pwd)/input_data:/input_data -it cita-vm-bench /usr/bin/python3 /scripts/merge.py
+docker run --env PYTHONIOENCODING=UTF-8 -v $(pwd)/../benchmark_results_data:/benchmark_results_data -v $(pwd)/evmraceresults:/evmraceresults -v $(pwd)/scripts:/scripts -it cita-vm-bench /usr/bin/python3 /scripts/merge.py
