+++ scripts/generate_report.py
@@ -1,167 +1,37 @@
+import sys
+
 import matplotlib.pyplot as plt
 import matplotlib.ticker as ticker
-
 import pandas as pd
 import numpy as np
 import durationpy
 import math
+
 from os.path import join
-from collections import defaultdict
-
-plt.style.use('ggplot')
-plt.rcParams.update({'figure.max_open_warning': 0})
-
-COLORS_DEFAULT = {'blue': '#348ABD', 'red': '#E24A33'}
-
+from plot_util import *
 from adjustText import adjust_text
 
 #############
 # CONSTANTS #
 #############
 
+if __name__ == "__main__":
 # csv result files directory name and file names
-CSV_RESULT_DIR = "../benchmark_results_data"
-WASM_RESULT_FILE = "standalone_wasm_results.csv"
-NATIVE_RESULT_FILE = "native_benchmarks.csv"
-SCOUT_RESULT_FILE = "scout_bignum_benchmarks.csv"
+    CSV_RESULT_DIR = "../benchmark_results_data"
+    WASM_RESULT_FILE = "standalone_wasm_results.csv"
+    NATIVE_RESULT_FILE = "native_benchmarks.csv"
+    SCOUT_RESULT_FILE = "scout_bignum_benchmarks.csv"
 
-IMG_OUTPUT_DIR = "../charts/"
 
 # To add a new engine, add the engine name used in the csv file to the list
-INTERPRETER_ENGINES = ['life', 'wagon', 'wasmi', 'wabt', 'v8-interpreter', 'wasm3', 'wamr']
-COMPILER_ENGINES = ['lifePolymerase', 'wasmtime', 'wavm', 'v8-liftoff', 'v8-turbofan', 'asmble']
+    INTERPRETER_ENGINES = ['life', 'wagon', 'wasmi', 'wabt', 'v8-interpreter', 'wasm3', 'wamr', 'fizzy']
+    COMPILER_ENGINES = ['lifePolymerase', 'wasmtime', 'wavm', 'v8-liftoff', 'v8-turbofan', 'asmble']
 
-wasm_vm_names = INTERPRETER_ENGINES + COMPILER_ENGINES
+    wasm_vm_names = INTERPRETER_ENGINES + COMPILER_ENGINES
 
-###################
-# HELPER FUNCTIONS
-###################
-def toMs(secs):
-    if secs < 0.001:
-        μs = secs * 1000000
-        return "{}μs".format(round(μs, 1))
-    if secs < 0.1:
-        ms = secs * 1000
-        return "{}ms".format(round(ms, 1))
-    elif secs < 1:
-        ms = secs * 1000
-        return "{}ms".format(int(round(ms, 0)))
-    else:
-        return "{}s".format(round(secs, 2))
-
-def labelBarHeights(ax, to_ms=True, small=False, to_factor=False, lower_y_bound=True):
-    labels = []
-    for p in ax.patches:
-        if p.get_height() == 0:
-            continue
-        label_val = p.get_height()
-        y_pos = p.get_y() + label_val
-        y_offset = 0
-        y_bound = ax.get_ybound()[1]
-        if y_pos > y_bound:
-            y_pos = y_bound
-        label_val = p.get_height()
-        y_pos = p.get_y() + label_val
-        y_offset = 0
-        y_bound = ax.get_ybound()[1]
-        if y_pos > y_bound:
-            y_pos = y_bound
-        if lower_y_bound:
-            if label_val < (0.03 * y_bound) and small is False:
-                # don't place labels near bottom axis, adjust_text cant deal
-                y_offset = 0.03 * y_bound
-        
-        if label_val > 1000:
-            label_val = int(label_val)
-        if to_ms:
-            label_val = toMs(label_val)
-        if to_factor:
-            if not isinstance(to_factor, int):
-                to_factor = 2
-            label_val = str(round(label_val, to_factor)) + "x"
-        """
-        txt_label = ax.annotate(toMs(label_val),
-                    (p.get_x()+p.get_width()/2.,
-                     y_pos),
-                    ha='center',
-                    va='center',
-                    xytext=(0, y_offset),
-                    textcoords='offset points',
-                    fontsize=10,
-                    weight="bold")
-        """
- 
-        txt_label = ax.text(
-                    x=p.get_x()+p.get_width()/2.,
-                    y=y_pos+y_offset,
-                    s=label_val,
-                    ha='center',
-                    va='center',
-                    #xytext=(0, y_offset),
-                    #textcoords='offset points',
-                    fontsize=10,
-                    weight="bold")
-
-        labels.append(txt_label)
-
-    return labels    
-
-
-def labelBarWidths(ax, to_ms=True, to_factor=False, to_kb=False, round_places=None):
-    for p in ax.patches:
-        label_val = p.get_width()
-        if label_val == 0:
-            continue
-        x_pos = p.get_x() + label_val
-        x_offset = -30
-        x_bound = ax.get_xbound()[1]
-        if (x_pos > x_bound):
-            x_pos = x_bound
-        if label_val < (0.15 * x_bound):
-            # if bar is too small to fit text, place next to bar
-            x_offset = 30
-        if label_val > 1000:
-            label_val = int(label_val)
-        if to_ms:
-            label_val = toMs(label_val)
-        if to_factor:
-            label_val = str(round(label_val, 2)) + "x"
-        if to_kb:
-            kbs = label_val / 1000.0
-            label_val = str(round(kbs, 1)) + " kB"
-        if round_places:
-            label_val = str(round(label_val, round_places))
-        ax.annotate(label_val,
-                    (x_pos,
-                     p.get_y()+p.get_height()/2.),
-                    ha='center',
-                    va='center',
-                    xytext=(x_offset, 0),
-                    textcoords='offset points',
-                    fontsize=10,
-                    weight="bold")
-
-def adjust_text_labels(labels, ax=None):
-    adjust_text(labels, ax=ax, autoalign='y', only_move={'points':'y', 'text':'y'})
-
-def read_results(file_name):
-    results = defaultdict(list)
-    csv_results = pd.read_csv(file_name)    
-    return csv_results
-                       
-def filterDfEngines(df_data, engines):
-    df_data_filtered = df_data.copy()
-    for fil in engines:
-        df_data_filtered = df_data_filtered[df_data_filtered['engine'] != fil]
-
-    return df_data_filtered
-
-####################################
-# print benchmark machine cpu type #
-####################################
-print('CPU Information')
-with open(join(CSV_RESULT_DIR, "cpuinfo.txt"), 'r') as cpuinfofile:
-    [print(line.rstrip()) for line in cpuinfofile.readlines()]
+    # print('CPU Information')
+    # with open(join(CSV_RESULT_DIR, "cpuinfo.txt"), 'r') as cpuinfofile:
+    #     [print(line.rstrip()) for line in cpuinfofile.readlines()]
 
 
 #########
@@ -169,400 +39,168 @@ with open(join(CSV_RESULT_DIR, "cpuinfo.txt"), 'r') as cpuinfofile:
 #########
 
 # Import scout engine results
-df_scout_data = read_results(join(CSV_RESULT_DIR, SCOUT_RESULT_FILE))
+    df_scout_data = read_results(join(CSV_RESULT_DIR, SCOUT_RESULT_FILE))
 
-# Functions for plotting a two engine comparison
-def plotThreeTestsGrouped(df_benches, three_tests, title="Title"):
-    df_1 = df_benches[df_benches['bench_name'].str.contains(three_tests[0])]
-    df_2 = df_benches[df_benches['bench_name'].str.contains(three_tests[1])]
-    df_3 = df_benches[df_benches['bench_name'].str.contains(three_tests[2])]
-    # to group by engine, pivot to `columns='enginge'`
-    df_1 = pd.pivot_table(df_1, values='exec_time', columns=['engine'], index=['bench_name'], aggfunc=np.mean)
-    df_2 = pd.pivot_table(df_2, values='exec_time', columns=['engine'], index=['bench_name'], aggfunc=np.mean)
-    df_3 = pd.pivot_table(df_3, values='exec_time', columns=['engine'], index=['bench_name'], aggfunc=np.mean)
-    filename = title.lower().replace(' ', '-')
-
-    plt.figure()
-    f, (ax1) = plt.subplots(1, 3, sharey=True, figsize=(13,6))
-
-    df_1.sort_values('wabt-with-superops').plot.bar(ax=ax1[0])
-    ax1[0].set_title(three_tests[0])
-    ax1[0].set(ylabel='seconds', xlabel='')
-
-    df_2.sort_values('wabt-with-superops').plot.bar(ax=ax1[1])
-    ax1[1].set_title(three_tests[1])
-    ax1[1].set(xlabel='')
-
-    df_3.sort_values('wabt-with-superops').plot.bar(ax=ax1[2])
-    ax1[2].set_title(three_tests[2])
-    ax1[2].set(xlabel='')
-
-    adjust_text_labels(labelBarHeights(ax1[0]), ax=ax1[0])
-    adjust_text_labels(labelBarHeights(ax1[1]), ax=ax1[1])
-    adjust_text_labels(labelBarHeights(ax1[2]), ax=ax1[2])
-
-    f.suptitle(title, fontsize=16, y=0.98)
-    print(f"plotTreeTestsGrouped: {filename}.png ")
-    plt.savefig(IMG_OUTPUT_DIR + filename + '.png')
 
 # Compare wabt-optimized against wabt-baseline on hash function benchmark
-plotThreeTestsGrouped(df_scout_data, ["blake2b_64", "blake2b_256", "blake2b_1024"], "blake2b C implementations compared")
-plotThreeTestsGrouped(df_scout_data, ["sha256_64", "sha256_256", "sha256_1024"], "sha256 C implementations compared")
-plotThreeTestsGrouped(df_scout_data, ["keccak256_64", "keccak256_256", "keccak256_1024"], "keccak256 C implementations compared")
 
-# Functions for plotting biturbo and bignum benchmarks
-# non-stacked bar chart, for only plotting exec time or total time
-def plotOneTest(df_data, suptitle="title", suptitle_pos=1.00, subtitle="subchart", subtitle_size='medium'):
-    filename = subtitle.strip().lower().replace(' ', '-')
-    plt.figure()
-    df_total_times = df_data[['exec_time', 'parse_time']].copy()
-    df_total_times.fillna(0, inplace=True)
-    df_total_times['time'] = df_total_times['exec_time'] + df_total_times['parse_time']
-    ax = df_total_times.sort_values('time')[['time']].plot.bar(figsize=(14,8), color=COLORS_DEFAULT['blue'])
-    # fontsize = {size in points, 'xx-small', 'x-small', 'small', 'medium', 'large', 'x-large', 'xx-large'}
-    ax.set_title(subtitle, fontsize=subtitle_size)
-    ax.set(ylabel="seconds", xlabel="")
-    ax.xaxis.set_tick_params(labelrotation=60, labelsize=12)
-    adjust_text_labels(labelBarHeights(ax))
-    plt.suptitle(suptitle, fontsize=18, y=suptitle_pos)
-    print(f"plotOneTest: {filename}.png")
-    plt.savefig(IMG_OUTPUT_DIR + filename + '.png')
-
-def plotScoutStackedTest(df_data, suptitle="title", suptitle_pos=1.00, subtitle="subchart", subtitle_size='medium'):
-    filename = subtitle.strip().lower().replace(' ', '-')
-    plt.figure()
-    # for some reason, exec_time needs to come first for the stacked chart to display correctly
-    ax = df_data[['exec_time', 'parse_time']].sort_values('exec_time').plot.bar(stacked=True, color=[COLORS_DEFAULT['blue'], COLORS_DEFAULT['red']], figsize=(10,10))
-    # fontsize = {size in points, 'xx-small', 'x-small', 'small', 'medium', 'large', 'x-large', 'xx-large'}
-    ax.set_title(subtitle, fontsize=subtitle_size)
-    ax.set(ylabel="seconds", xlabel="")
-    df_total_times = df_data[['exec_time', 'parse_time']].copy()
-    df_total_times['total_time'] = df_total_times['exec_time'] + df_total_times['parse_time']
-    ax.xaxis.set_tick_params(labelrotation=60, labelsize=12)
-    adjust_text_labels(labelBarHeights(ax, lower_y_bound=False))
-    ax.legend(labels=["execution time", "startup time"])
-    plt.suptitle(suptitle, fontsize=18, y=suptitle_pos)
-    print(f"plotScoutStackedTest: {filename}.png")
-    plt.savefig(IMG_OUTPUT_DIR + filename + '.png')
 
 # Plot biturbo benchmark:
 # biturbo-token-eth1-mainnet-stateless-block-hexary-trie-keccac256-multiproof
-df_scout_biturbo_token = df_scout_data[df_scout_data['bench_name'] == 'biturbo-token-eth1-mainnet-stateless-block-hexary-trie-keccak256-multiproof']
+    df_scout_biturbo_token = df_scout_data[df_scout_data['bench_name'] == 'biturbo-token-eth1-mainnet-stateless-block-hexary-trie-keccak256-multiproof']
 
-df_scout_means_biturbo_token = df_scout_biturbo_token.groupby(['engine']).mean()
-df_scout_fast_biturbo_token = filterDfEngines(df_scout_biturbo_token, ['v8-interpreter'])
-df_scout_fast_biturbo_means = df_scout_fast_biturbo_token.groupby(['engine']).mean()
+    df_scout_means_biturbo_token = df_scout_biturbo_token.groupby(['engine']).mean()
+    df_scout_fast_biturbo_token = filterDfEngines(df_scout_biturbo_token, ['v8-interpreter'])
+    df_scout_fast_biturbo_means = df_scout_fast_biturbo_token.groupby(['engine']).mean()
 
-plotOneTest(df_scout_means_biturbo_token,
-            suptitle="\nbiturbo token - all scout engines",
-            suptitle_pos=1.05,
-            subtitle="biturbo-token-eth1-mainnet-stateless-block-hexary-trie-keccak256-multiproof\n",
-            subtitle_size='xx-large')
 
-plotScoutStackedTest(df_scout_fast_biturbo_means,
-            suptitle="\nbiturbo token - fast scout engines - wasm compilers vs interpreters (v8 vs wabt)",
-            suptitle_pos=1.03,
-            subtitle="biturbo-token-eth1-mainnet-stateless-block-hexary-trie-keccak256-multiproof\n",
-            subtitle_size='xx-large')
 
 # Plot bignum benchmark:
 # ecrecover-eth1-txns-websnark-secp256k1-verify-72-sigs
-df_scout_ecrecover_websnark_secp256k1 = df_scout_data[df_scout_data['bench_name'] == 'ecrecover-eth1-txns-websnark-secp256k1-verify-72-sigs']
+    df_scout_ecrecover_websnark_secp256k1 = df_scout_data[df_scout_data['bench_name'] == 'ecrecover-eth1-txns-websnark-secp256k1-verify-72-sigs']
 
-df_scout_means_ecrecover_websnark_secp256k1 = df_scout_ecrecover_websnark_secp256k1.groupby(['engine']).mean()
-df_scout_fast_ecrecover = filterDfEngines(df_scout_ecrecover_websnark_secp256k1,
-    ['v8-interpreter', 'scoutcpp-wabt-no-bignums', 'wabt-no-bignums'])
-df_scout_fast_ecrecover_means = df_scout_fast_ecrecover.groupby(['engine']).mean()
+    df_scout_means_ecrecover_websnark_secp256k1 = df_scout_ecrecover_websnark_secp256k1.groupby(['engine']).mean()
+    df_scout_fast_ecrecover = filterDfEngines(df_scout_ecrecover_websnark_secp256k1,
+        ['v8-interpreter', 'scoutcpp-wabt-no-bignums', 'wabt-no-bignums'])
+    df_scout_fast_ecrecover_means = df_scout_fast_ecrecover.groupby(['engine']).mean()
 
-plotOneTest(df_scout_means_ecrecover_websnark_secp256k1,
-            suptitle="websnark-secp256k1-sig-verify - all Scout engines",
-            suptitle_pos=1.0,
-            subtitle="ecrecover-eth1-txns-websnark-secp256k1-verify-72-sigs\n",
-            subtitle_size='xx-large')
 
-plotScoutStackedTest(df_scout_fast_ecrecover_means,
-            #suptitle="websnark-secp256k1-sig-verify - fast Scout engines - compilers (v8) vs interpreters (wabt)",
-            suptitle="compiler engines - optimizing (v8-turbofan) and single-pass (v8-liftoff) \n vs. \n interpreter engine (wabt) with bignum host funcs",
-            suptitle_pos=1.07,
-            subtitle="ecrecover-eth1-txns-websnark-secp256k1-verify-72-sigs\n",
-            subtitle_size='xx-large')
 
 # Plot bignum benhmark:
 # ecpairing-zkrollup-websnark-bn128-two-pairings
-df_scout_ecpairing_zkrollup_bn128_pairings = df_scout_data[df_scout_data['bench_name'] == 'ecpairing-zkrollup-websnark-bn128-two-pairings']
+    df_scout_ecpairing_zkrollup_bn128_pairings = df_scout_data[df_scout_data['bench_name'] == 'ecpairing-zkrollup-websnark-bn128-two-pairings']
 
-df_scout_means_ecpairing_zkrollup = df_scout_ecpairing_zkrollup_bn128_pairings.groupby(['engine']).mean()
-df_scout_fast_ecpairing_zkrollup = filterDfEngines(df_scout_ecpairing_zkrollup_bn128_pairings,
-    ['v8-interpreter', 'scoutcpp-wabt-no-bignums', 'wabt-no-bignums'])
-df_scout_fast_means_ecpairing_zkrollup = df_scout_fast_ecpairing_zkrollup.groupby(['engine']).mean()
+    df_scout_means_ecpairing_zkrollup = df_scout_ecpairing_zkrollup_bn128_pairings.groupby(['engine']).mean()
+    df_scout_fast_ecpairing_zkrollup = filterDfEngines(df_scout_ecpairing_zkrollup_bn128_pairings,
+        ['v8-interpreter', 'scoutcpp-wabt-no-bignums', 'wabt-no-bignums'])
+    df_scout_fast_means_ecpairing_zkrollup = df_scout_fast_ecpairing_zkrollup.groupby(['engine']).mean()
 
-plotOneTest(df_scout_means_ecpairing_zkrollup,
-            suptitle="websnark-bn128-pairings - all Scout engines",
-            suptitle_pos=1.02,
-            subtitle="ecpairing-zkrollup-websnark-bn128-two-pairings\n",
-            subtitle_size='xx-large')
 
-plotScoutStackedTest(df_scout_fast_means_ecpairing_zkrollup,
-            suptitle="compiler engines - optimizing (v8-turbofan) and single-pass (v8-liftoff) \n vs. \n interpreter engine (wabt) with bignum host funcs",
-            suptitle_pos=1.07,
-            subtitle="ecpairing-zkrollup-websnark-bn128-two-pairings\n",
-            subtitle_size='xx-large')
 
 # Plot bignum benchmark:
 # daiquiri-zkmixer-websnark-bn128-groth16-four-pairings-and-mimc
-df_scout_daiquiri_zkmixer = df_scout_data[df_scout_data['bench_name'] == 'daiquiri-zkmixer-websnark-bn128-groth16-four-pairings-and-mimc']
+    df_scout_daiquiri_zkmixer = df_scout_data[df_scout_data['bench_name'] == 'daiquiri-zkmixer-websnark-bn128-groth16-four-pairings-and-mimc']
 
-df_scout_means_daiquiri_zkmixer = df_scout_daiquiri_zkmixer.groupby(['engine']).mean()
-df_scout_fast_daiquiri_zkmixer = filterDfEngines(df_scout_daiquiri_zkmixer,
-    ['v8-interpreter', 'scoutcpp-wabt-no-bignums', 'wabt-no-bignums'])
-df_scout_fast_means_daiquiri_zkmixer = df_scout_fast_daiquiri_zkmixer.groupby(['engine']).mean()
+    df_scout_means_daiquiri_zkmixer = df_scout_daiquiri_zkmixer.groupby(['engine']).mean()
+    df_scout_fast_daiquiri_zkmixer = filterDfEngines(df_scout_daiquiri_zkmixer,
+        ['v8-interpreter', 'scoutcpp-wabt-no-bignums', 'wabt-no-bignums'])
+    df_scout_fast_means_daiquiri_zkmixer = df_scout_fast_daiquiri_zkmixer.groupby(['engine']).mean()
 
-plotOneTest(df_scout_means_daiquiri_zkmixer,
-            suptitle="daiquiri-zkmixer - all Scout engines",
-            suptitle_pos=1.02,
-            subtitle="daiquiri-zkmixer-websnark-bn128-groth16-four-pairings-and-mimc\n",
-            subtitle_size='xx-large')
 
-plotScoutStackedTest(df_scout_fast_means_daiquiri_zkmixer,
-            suptitle="wasm compilers - optimizing (v8-turbofan) and single-pass (v8-liftoff) \n vs. \n wasm interpreter (wabt) with bignum host funcs",
-            suptitle_pos=1.07,
-            subtitle="daiquiri-zkmixer-websnark-bn128-groth16-four-pairings-and-mimc\n",
-            subtitle_size='xx-large')
 
 ###########################
 # Wasm Engines Benchmarks #
 ###########################
 
 # Import wasm results
-df_wasm = read_results(join(CSV_RESULT_DIR, WASM_RESULT_FILE))
+    df_wasm = read_results(join(CSV_RESULT_DIR, WASM_RESULT_FILE))
 
 # Import native results
-df_native_raw = read_results(join(CSV_RESULT_DIR, NATIVE_RESULT_FILE))
+    df_native_raw = read_results(join(CSV_RESULT_DIR, NATIVE_RESULT_FILE))
 
 
-native_results = {}
-for i in range(len(df_native_raw)):
-    test_name = df_native_raw['test_name'][i]
-    native_results[test_name] = df_native_raw['elapsed_times'][i].split(', ')
+    native_results = {}
+    for i in range(len(df_native_raw)):
+        test_name = df_native_raw['test_name'][i]
+        native_results[test_name] = df_native_raw['elapsed_times'][i].split(', ')
 
-test_names = df_native_raw['test_name'].tolist()
+    test_names = df_native_raw['test_name'].tolist()
 
-native_times = {}
-for test in test_names:
-    avg = np.mean([float(t) for t in native_results[test]])
-    native_times[test] = avg
-    
-df_native = pd.DataFrame.from_dict(native_times, orient='index', columns=['elapsed_time'])
-df_native['engine'] = 'rust-native'
+    native_times = {}
+    for test in test_names:
+        avg = np.mean([float(t) for t in native_results[test]])
+        native_times[test] = avg
+        
+    df_native = pd.DataFrame.from_dict(native_times, orient='index', columns=['elapsed_time'])
+    df_native['engine'] = 'rust-native'
 
 # Sorted list of test names for plotting inputs in order
-blake2b_test_names = [name for name in df_wasm['test_name'].unique().tolist() if 'blake2b' in name]
-blake2b_test_names.sort()
-blake2b_test_names_desc = blake2b_test_names.copy()
-blake2b_test_names_desc.reverse()
+    blake2b_test_names = [name for name in df_wasm['test_name'].unique().tolist() if 'blake2b' in name]
+    blake2b_test_names.sort()
+    blake2b_test_names_desc = blake2b_test_names.copy()
+    blake2b_test_names_desc.reverse()
 
 
-sha1_test_names = [name for name in df_wasm['test_name'].unique().tolist() if 'sha1' in name]
-sha1_test_names.sort()
-sha1_test_names_desc = sha1_test_names.copy()
-sha1_test_names_desc.reverse()
+    sha1_test_names = [name for name in df_wasm['test_name'].unique().tolist() if 'sha1' in name]
+    sha1_test_names.sort()
+    sha1_test_names_desc = sha1_test_names.copy()
+    sha1_test_names_desc.reverse()
 
 
-bn128_mul_test_names = ['bn128_mul-cdetrio11', 'bn128_mul-chfast2', 'bn128_mul-chfast1']
-bn128_add_test_names = ['bn128_add-cdetrio11', 'bn128_add-cdetrio14', 'bn128_add-cdetrio10']
-bn128_pairing_test_names = ['bn128_pairing-ten_point_match_1', 'bn128_pairing-two_point_match_2', 'bn128_pairing-one_point']
+    bn128_mul_test_names = ['bn128_mul-cdetrio11', 'bn128_mul-chfast2', 'bn128_mul-chfast1']
+    bn128_add_test_names = ['bn128_add-cdetrio11', 'bn128_add-cdetrio14', 'bn128_add-cdetrio10']
+    bn128_pairing_test_names = ['bn128_pairing-ten_point_match_1', 'bn128_pairing-two_point_match_2', 'bn128_pairing-one_point']
 
-bn128_mul_test_names.reverse()
-bn128_add_test_names.reverse()
-bn128_pairing_test_names.reverse()
+    bn128_mul_test_names.reverse()
+    bn128_add_test_names.reverse()
+    bn128_pairing_test_names.reverse()
 
-bn128_mul_test_names_desc = ['bn128_mul-cdetrio11', 'bn128_mul-chfast2', 'bn128_mul-chfast1']
-bn128_add_test_names_desc = ['bn128_add-cdetrio11', 'bn128_add-cdetrio14', 'bn128_add-cdetrio10']
-bn128_pairing_test_names_desc = ['bn128_pairing-ten_point_match_1', 'bn128_pairing-two_point_match_2', 'bn128_pairing-one_point']
+    bn128_mul_test_names_desc = ['bn128_mul-cdetrio11', 'bn128_mul-chfast2', 'bn128_mul-chfast1']
+    bn128_add_test_names_desc = ['bn128_add-cdetrio11', 'bn128_add-cdetrio14', 'bn128_add-cdetrio10']
+    bn128_pairing_test_names_desc = ['bn128_pairing-ten_point_match_1', 'bn128_pairing-two_point_match_2', 'bn128_pairing-one_point']
 
-bls12_test_names = ['bls12-381-aggreg-32-pubkeys-verify-sig', 'bls12-381-aggreg-64-pubkeys-verify-sig', 'bls12-381-aggreg-128-pubkeys-verify-sig']
-bls12_test_names_desc = bls12_test_names.copy()
-bls12_test_names_desc.reverse()
+    bls12_test_names = ['bls12-381-aggreg-32-pubkeys-verify-sig', 'bls12-381-aggreg-64-pubkeys-verify-sig', 'bls12-381-aggreg-128-pubkeys-verify-sig']
+    bls12_test_names_desc = bls12_test_names.copy()
+    bls12_test_names_desc.reverse()
 
-modexp_test_names = [name for name in df_wasm['test_name'].unique().tolist() if 'modexp' in name]
-modexp_test_names.sort()
-modexp_test_names_desc = modexp_test_names.copy()
-modexp_test_names_desc.reverse()
+    modexp_test_names = [name for name in df_wasm['test_name'].unique().tolist() if 'modexp' in name]
+    modexp_test_names.sort()
+    modexp_test_names_desc = modexp_test_names.copy()
+    modexp_test_names_desc.reverse()
 
-all_tests = sha1_test_names + blake2b_test_names + modexp_test_names + ['ed25519-verify-ten-inputs'] + bls12_test_names \
-    + bn128_add_test_names + bn128_mul_test_names + bn128_pairing_test_names
+    all_tests = sha1_test_names + blake2b_test_names + modexp_test_names + ['ed25519-verify-ten-inputs'] + bls12_test_names \
+        + bn128_add_test_names + bn128_mul_test_names + bn128_pairing_test_names
 
 # Prepare interpreter dataframe
-all_interp_test_names = []
+    all_interp_test_names = []
 
-interp_results_for_df = []
-for engine in INTERPRETER_ENGINES:
-    df_engine = df_wasm[df_wasm['engine'] == engine]
-    df_means = df_engine.groupby(['test_name']).mean()
-    test_names = df_engine['test_name'].unique().tolist()
-    all_interp_test_names.extend(test_names)
-    for test_name in test_names:
-        interp_results_for_df.append([engine, test_name] + df_means.loc[test_name].tolist())
+    interp_results_for_df = []
+    for engine in INTERPRETER_ENGINES:
+        df_engine = df_wasm[df_wasm['engine'] == engine]
+        df_means = df_engine.groupby(['test_name']).mean()
+        test_names = df_engine['test_name'].unique().tolist()
+        all_interp_test_names.extend(test_names)
+        for test_name in test_names:
+            interp_results_for_df.append([engine, test_name] + df_means.loc[test_name].tolist())
 
-all_interp_test_names = set(all_interp_test_names)
+    all_interp_test_names = set(all_interp_test_names)
 
-df_interp = pd.DataFrame(interp_results_for_df)
-df_interp.columns = ['engine', 'test_name', 'elapsed_time', 'parse_time', 'exec_time']
-df_interp.set_index('engine', inplace=True)
-df_interp['total_time'] = df_interp['parse_time'] + df_interp['exec_time']
+    df_interp = pd.DataFrame(interp_results_for_df)
+    df_interp.columns = ['engine', 'test_name', 'elapsed_time', 'parse_time', 'exec_time']
+    df_interp.set_index('engine', inplace=True)
+    df_interp['total_time'] = df_interp['parse_time'] + df_interp['exec_time']
 
-# Interpreter engines compared
-def plotInterpThreeTests(df_testdata, three_names, title="Title", filter_engines=None):
-    filename = title.lower().replace(' ', '-')
-    if filter_engines is not None:
-        df_testdata = filterDfEngines(df_testdata.reset_index(), filter_engines)
-        df_testdata.set_index('engine', inplace=True)
-    df_1 = df_testdata[df_testdata['test_name'] == three_names[0]]
-    df_2 = df_testdata[df_testdata['test_name'] == three_names[1]]
-    df_3 = df_testdata[df_testdata['test_name'] == three_names[2]]
-    plt.figure()
-    f, (ax1) = plt.subplots(1, 3, sharey=True, figsize=(16,7))
-    df_1.sort_values('total_time')[['parse_time', 'exec_time']].plot.bar(ax=ax1[0], stacked=True)
-    ax1[0].set_title(three_names[0])
-    ax1[0].set(xlabel='')
-    df_2.sort_values('total_time')[['parse_time', 'exec_time']].plot.bar(ax=ax1[1], stacked=True)
-    ax1[1].set_title(three_names[1])
-    ax1[1].set(xlabel='')
-    df_3.sort_values('total_time')[['parse_time', 'exec_time']].plot.bar(ax=ax1[2], stacked=True)
-    ax1[2].set_title(three_names[2])
-    ax1[2].set(xlabel='')
-    adjust_text_labels(labelBarHeights(ax1[0]), ax=ax1[0])
-    adjust_text_labels(labelBarHeights(ax1[1]), ax=ax1[1])
-    adjust_text_labels(labelBarHeights(ax1[2]), ax=ax1[2])
-    ax1[0].legend(labels=["startup time", "execution time"])
-    ax1[1].legend(labels=["startup time", "execution time"])
-    ax1[2].legend(labels=["startup time", "execution time"])
 
-    ax1[1].tick_params(axis='y', left=True, labelleft=True)
-    ax1[2].tick_params(axis='y', left=True, labelleft=True)
-    
-    scale_y = 1e-3
-    ticks_y = ticker.FuncFormatter(lambda y, pos: '{0:g}'.format(y/scale_y))
-    ax1[1].yaxis.set_major_formatter(ticks_y)
-    
-    ax1[0].set(ylabel='milliseconds')
-    
-    f.suptitle(title, fontsize=16, y=0.98)
-    print(f"plotInterpThreeTests: {filename}.png")
-    plt.savefig(IMG_OUTPUT_DIR + filename + '.png')
 
-plotInterpThreeTests(df_interp, blake2b_test_names, title="wasm interpreters compared - blake2b")
 
-# Interpreters - Execution time
-def plotThreeTestsExecTime(df_testdata, three_names, title="Title", filter_engines=None):
-    filename = title.lower().replace(' ', '-')
-    if filter_engines is not None:
-        df_testdata = filterDfEngines(df_testdata.reset_index(), filter_engines)
-        df_testdata.set_index('engine', inplace=True)
-    df_1 = df_testdata[df_testdata['test_name'] == three_names[0]]
-    df_2 = df_testdata[df_testdata['test_name'] == three_names[1]]
-    df_3 = df_testdata[df_testdata['test_name'] == three_names[2]]
-    plt.figure()
-    f, (ax1) = plt.subplots(1, 3, sharey=True, figsize=(16,7))
-    df_1.sort_values('exec_time')[['exec_time']].plot.bar(ax=ax1[0], color=COLORS_DEFAULT['blue'])
-    ax1[0].set_title(three_names[0])
-    ax1[0].set(xlabel='')
-    df_2.sort_values('exec_time')[['exec_time']].plot.bar(ax=ax1[1], color=COLORS_DEFAULT['blue'])
-    ax1[1].set_title(three_names[1])
-    ax1[1].set(xlabel='')
-    df_3.sort_values('exec_time')[['exec_time']].plot.bar(ax=ax1[2], color=COLORS_DEFAULT['blue'])
-    ax1[2].set_title(three_names[2])
-    ax1[2].set(xlabel='')
-    adjust_text_labels(labelBarHeights(ax1[0]), ax=ax1[0])
-    adjust_text_labels(labelBarHeights(ax1[1]), ax=ax1[1])
-    adjust_text_labels(labelBarHeights(ax1[2]), ax=ax1[2])
-    ax1[0].legend(labels=["execution time"])
-    ax1[1].legend(labels=["execution time"])
-    ax1[2].legend(labels=["execution time"])
 
-    ax1[1].tick_params(axis='y', left=True, labelleft=True)
-    ax1[2].tick_params(axis='y', left=True, labelleft=True)
-   
-    scale_y = 1e-3
-    ticks_y = ticker.FuncFormatter(lambda y, pos: '{0:g}'.format(y/scale_y))
-    ax1[1].yaxis.set_major_formatter(ticks_y)
-    
-    ax1[0].set(ylabel='milliseconds')
-    
-    f.suptitle(title, fontsize=16, y=0.98)
-    print(f"plotThreeTestsExecTime: {filename}.png")
-    plt.savefig(IMG_OUTPUT_DIR + filename + '.png')
 
-plotThreeTestsExecTime(df_interp, blake2b_test_names, title="wasm interpreters compared - blake2b")
+
 
 # Chart of only the fast interpreters
-plotInterpThreeTests(df_interp,
-                     blake2b_test_names,
-                     title="wasm interpreters compared - blake2b",
-                     filter_engines=["life", "wagon", "v8-interpreter"])
 
-plotInterpThreeTests(df_interp,
-                     modexp_test_names,
-                     title="wasm interpreters compared - modexp",
-                     filter_engines=["life", "wagon", "v8-interpreter"])
-
-plotInterpThreeTests(df_interp,
-                     bn128_add_test_names,
-                     title="wasm interpreters compared - bn128_add",
-                     filter_engines=["life", "wagon", "v8-interpreter"])
-
-plotThreeTestsExecTime(df_interp,
-                       blake2b_test_names,
-                       title="wasm interpreters compared - blake2b",
-                       filter_engines=["life", "wagon", "v8-interpreter"])
-
-plotThreeTestsExecTime(df_interp,
-                       bls12_test_names,
-                       title="wasm interpreters compared - bls12",
-                       filter_engines=["life", "wagon", "v8-interpreter"])
 
 ####################
 # Compiler Results #
 ####################
 
 # Prepare compilers dataframe
-compiler_results_for_df = []
-for engine in COMPILER_ENGINES:
-    df_engine = df_wasm[df_wasm['engine'] == engine]
-    df_means = df_engine.groupby(['test_name']).mean()
-    test_names = df_engine['test_name'].unique().tolist()
-    for test_name in test_names:
-        compiler_results_for_df.append([engine, test_name] + df_means.loc[test_name].tolist())
+    compiler_results_for_df = []
+    for engine in COMPILER_ENGINES:
+        df_engine = df_wasm[df_wasm['engine'] == engine]
+        df_means = df_engine.groupby(['test_name']).mean()
+        test_names = df_engine['test_name'].unique().tolist()
+        for test_name in test_names:
+            compiler_results_for_df.append([engine, test_name] + df_means.loc[test_name].tolist())
 
-df_compiler = pd.DataFrame(compiler_results_for_df)
-df_compiler.columns = ['engine', 'test_name', 'elapsed_time', 'compile_time', 'exec_time']
-df_compiler.set_index('engine', inplace=True)
-df_compiler['total_time'] = df_compiler['compile_time'] + df_compiler['exec_time']
+    df_compiler = pd.DataFrame(compiler_results_for_df)
+    df_compiler.columns = ['engine', 'test_name', 'elapsed_time', 'compile_time', 'exec_time']
+    df_compiler.set_index('engine', inplace=True)
+    df_compiler['total_time'] = df_compiler['compile_time'] + df_compiler['exec_time']
 
-# Compiler function
-# Note: Fix scale for v8-turbofan execution time
-def plotCompilerStackedOneTest(df_benchdata, test_name, native=False):
-    filename = test_name
-    if native: filename += '-native'
-    df_1 = df_benchdata[df_benchdata['test_name'] == test_name]
-    # for some reason, exec_time needs to come first for the stacked chart to display correctly
-    ax = df_1[['exec_time', 'compile_time']].sort_values('exec_time').plot.bar(stacked=True, color=[COLORS_DEFAULT['blue'], COLORS_DEFAULT['red']], figsize=(10,10))
-    ax.set_title("Compilers - {}".format(test_name))
-    df_total_times = df_1[['exec_time', 'compile_time']].copy()
-    df_total_times['total_time'] = df_total_times['exec_time'] + df_total_times['compile_time']
-    ymax = max(df_total_times[['total_time']].max()) * 1.3 # 30% larger for padding between top of bars and top of graph
-    ymin = min([x for x in df_1[['exec_time', 'compile_time']].min().tolist() if x > 0])
-    ymin = ymin * 0.8 # 0.8 to get a number 20% smaller, fix to make bar appear for smallest exec time
-    ax.set_ylim(ymin, ymax)
-    ax.set_yscale("log")
-    plt.ylabel("seconds (log scale)")
-    adjust_text_labels(labelBarHeights(ax, lower_y_bound=False))
-    ax.legend(labels=["execution time", "compile time"])
-    print(f"plotCompilerStackedOneTest: {filename}.png")
-    plt.savefig(IMG_OUTPUT_DIR + filename + '.png')
 
-plotCompilerStackedOneTest(df_compiler, blake2b_test_names[2])
-plotCompilerStackedOneTest(df_compiler, "bls12-381-aggreg-128-pubkeys-verify-sig")
+
+
 
 
 # Add rust-native to compiler engines chart
@@ -571,101 +209,278 @@ plotCompilerStackedOneTest(df_compiler, "bls12-381-aggreg-128-pubkeys-verify-sig
 # both dataframes must have same columns to merge them:
 # engine, test_name, elapsed_time, compile_time, exec_time, total_time
 
-df_native_merge = df_native.copy()
-df_native_merge.reset_index(inplace=True)
-df_native_merge.columns = ['test_name', 'elapsed_time', 'engine']
-df_native_merge['compile_time'] = 0
-df_native_merge['exec_time'] = df_native_merge['elapsed_time']
-df_native_merge['total_time'] = df_native_merge['elapsed_time']
-df_native_and_compile = pd.concat([df_compiler.reset_index(), df_native_merge], sort=False)
-df_native_and_compile.reset_index(drop=True, inplace=True)
-df_native_and_compile.set_index('engine', inplace=True)
+    df_native_merge = df_native.copy()
+    df_native_merge.reset_index(inplace=True)
+    df_native_merge.columns = ['test_name', 'elapsed_time', 'engine']
+    df_native_merge['compile_time'] = 0
+    df_native_merge['exec_time'] = df_native_merge['elapsed_time']
+    df_native_merge['total_time'] = df_native_merge['elapsed_time']
+    df_native_and_compile = pd.concat([df_compiler.reset_index(), df_native_merge], sort=False)
+    df_native_and_compile.reset_index(drop=True, inplace=True)
+    df_native_and_compile.set_index('engine', inplace=True)
 
-plotCompilerStackedOneTest(df_native_and_compile, "bls12-381-aggreg-128-pubkeys-verify-sig", native=True)
+
 
 ###################################
 # Interpreter vs Compiler speedup #
 ###################################
 
 # merge df_compiler and df_interp
-df_interp_merge = df_interp.copy()
-df_interp_merge.columns = ['test_name', 'elapsed_time', 'compile_time', 'exec_time', 'total_time']
-df_interp_and_compile = pd.concat([df_interp_merge, df_compiler])
-df_interp_and_compile = df_interp_and_compile.reset_index()
+    df_interp_merge = df_interp.copy()
+    df_interp_merge.columns = ['test_name', 'elapsed_time', 'compile_time', 'exec_time', 'total_time']
+    df_interp_and_compile = pd.concat([df_interp_merge, df_compiler])
+    df_interp_and_compile = df_interp_and_compile.reset_index()
 
-def add_engine_ratio_col(df_two_engines, interp_engine, compiler_engine):
-    df_two = df_two_engines.copy(deep=True)
-    df_interp = df_two[df_two['engine'] == interp_engine].copy()
-    df_compile = df_two[df_two['engine'] == compiler_engine]
-    for index, row in df_interp.iterrows():
-        test_name = row['test_name']
-        row_compiler = df_compile[df_compile['test_name'] == test_name]
-        if len(row_compiler) != 1:
-            print("ERROR. SHOULD HAVE 1 MATCH ONLY")
-            break
-        compiler_exec_time = row_compiler.iloc[0]['exec_time']
-        exec_ratio = row['exec_time'] / compiler_exec_time
-        df_interp.at[index, 'compiler_exec_time'] = compiler_exec_time
-        df_interp.at[index, 'interp_slowdown'] = exec_ratio
 
-    return df_interp
 
-df_wabt_v8liftoff = add_engine_ratio_col(df_interp_and_compile, "wabt", "v8-liftoff")
+    df_wabt_v8liftoff = add_engine_ratio_col(df_interp_and_compile, "wabt", "v8-liftoff")
+    df_fizzy_v8liftoff = add_engine_ratio_col(df_interp_and_compile, "fizzy", "v8-liftoff")
+    df_wasm3_v8liftoff = add_engine_ratio_col(df_interp_and_compile, "wasm3", "v8-liftoff")
+    df_wasmi_wavm = add_engine_ratio_col(df_interp_and_compile, "wasmi", "wavm")
 
-def plotCompilerSpeedup(df_compiler_speedup, interp_name="interp", compiler_name="compiler"):
-    filename = 'compiler-speedup-' + interp_name + '-' + compiler_name
-    df_is = df_compiler_speedup.copy()
-    plot_tests = df_is['test_name'].tolist()
-    test_index = all_tests.copy()
-    for test_name in all_tests:
-        if test_name not in plot_tests:
-            test_index.remove(test_name)
- 
-    df_is = df_is.set_index('test_name')
-    df_is = df_is.reindex(test_index)
-    plt.figure()
-    ax = df_is['interp_slowdown'].plot.bar(figsize=(13,7), color=COLORS_DEFAULT['blue'])
-    ax.set_title("{} (Interpreter) vs {} (Compiler) - speedup".format(interp_name, compiler_name))
-    ax.set(ylabel="compiler speedup (log scale)", xlabel="")
 
-    y_max = ax.get_ylim()[1]
-    y_max = y_max * 5 # padding between top of bars and legend
-    ax.set_yscale('log')
-    ax.set_ylim(0.1, y_max) # adjustment to move horizontal line higher on chart
 
-    adjust_text_labels(labelBarHeights(ax, to_ms=False, to_factor=1, lower_y_bound=False))
 
-    plt.axhline(y=1.0, linewidth=2, ls='--', color='r')
 
-    handles, _ = ax.get_legend_handles_labels()
+    plotThreeTestsGrouped(df_scout_data, ["blake2b_64", "blake2b_256", "blake2b_1024"], "blake2b C implementations compared")
+    plotThreeTestsGrouped(df_scout_data, ["sha256_64", "sha256_256", "sha256_1024"], "sha256 C implementations compared")
+    plotThreeTestsGrouped(df_scout_data, ["keccak256_64", "keccak256_256", "keccak256_1024"], "keccak256 C implementations compared")
+    plotOneTest(df_scout_means_biturbo_token,
+                suptitle="\nbiturbo token - all scout engines",
+                suptitle_pos=1.05,
+                subtitle="biturbo-token-eth1-mainnet-stateless-block-hexary-trie-keccak256-multiproof\n",
+                subtitle_size='xx-large')
 
-    plt.legend(labels=["compiler engine is faster (ratio > 1x)", "compiler speedup \n (ratio: interp_exec_time/compiler_exec_time)"])
+    plotScoutStackedTest(df_scout_fast_biturbo_means,
+                suptitle="\nbiturbo token - fast scout engines - wasm compilers vs interpreters (v8 vs wabt)",
+                suptitle_pos=1.03,
+                subtitle="biturbo-token-eth1-mainnet-stateless-block-hexary-trie-keccak256-multiproof\n",
+                subtitle_size='xx-large')
+    plotOneTest(df_scout_means_ecrecover_websnark_secp256k1,
+                suptitle="websnark-secp256k1-sig-verify - all Scout engines",
+                suptitle_pos=1.0,
+                subtitle="ecrecover-eth1-txns-websnark-secp256k1-verify-72-sigs\n",
+                subtitle_size='xx-large')
 
-    #f.suptitle(title, fontsize=16, y=0.95)
-    print(f"plotCompilerSpeedup: {filename}.png")
-    plt.savefig(IMG_OUTPUT_DIR + filename + '.png')
+    plotScoutStackedTest(df_scout_fast_ecrecover_means,
+                #suptitle="websnark-secp256k1-sig-verify - fast Scout engines - compilers (v8) vs interpreters (wabt)",
+                suptitle="compiler engines - optimizing (v8-turbofan) and single-pass (v8-liftoff) \n vs. \n interpreter engine (wabt) with bignum host funcs",
+                suptitle_pos=1.07,
+                subtitle="ecrecover-eth1-txns-websnark-secp256k1-verify-72-sigs\n",
+                subtitle_size='xx-large')
+    plotOneTest(df_scout_means_ecpairing_zkrollup,
+                suptitle="websnark-bn128-pairings - all Scout engines",
+                suptitle_pos=1.02,
+                subtitle="ecpairing-zkrollup-websnark-bn128-two-pairings\n",
+                subtitle_size='xx-large')
 
-plotCompilerSpeedup(df_wabt_v8liftoff, interp_name="wabt", compiler_name="v8-liftoff")
+    plotScoutStackedTest(df_scout_fast_means_ecpairing_zkrollup,
+                suptitle="compiler engines - optimizing (v8-turbofan) and single-pass (v8-liftoff) \n vs. \n interpreter engine (wabt) with bignum host funcs",
+                suptitle_pos=1.07,
+                subtitle="ecpairing-zkrollup-websnark-bn128-two-pairings\n",
+                subtitle_size='xx-large')
+    plotOneTest(df_scout_means_daiquiri_zkmixer,
+                suptitle="daiquiri-zkmixer - all Scout engines",
+                suptitle_pos=1.02,
+                subtitle="daiquiri-zkmixer-websnark-bn128-groth16-four-pairings-and-mimc\n",
+                subtitle_size='xx-large')
 
-########################################################
-# All precompiles compared (are interpreters feasible? #
-########################################################
-def plotInterpOneEngine(df_benchdata, engine, test_names, title=""):
-    filename = engine + '-all-precompiles-compared'
-    df_bench = df_benchdata.loc[engine][df_benchdata.loc[engine]['test_name'].isin(test_names)]
-    df_bench.set_index('test_name', inplace=True)
-    df_bench = df_bench.reindex(test_names)
-    plt.figure()
-    ax = df_bench[['parse_time', 'exec_time']].plot.bar(stacked=True, figsize=(15,7), color=[COLORS_DEFAULT['red'], COLORS_DEFAULT['blue']])
-    plt.ylabel("seconds")
-    ax.set(xlabel='')
-    # doing get_legend_handles_labels() so we can control the order of the labels in the legend 
-    handles, labels = ax.get_legend_handles_labels()
-    plt.legend([handles[1], handles[0]], ['execution time (top bar/label)', 'startup time (bottom bar/label - bar not visible for times < 20.0ms)'])
-    adjust_text_labels(labelBarHeights(ax))
-    plt.suptitle(title, fontsize=16, y=0.95)
-    print(f"plotInterpOneEngine: {filename}.png")
-    plt.savefig(IMG_OUTPUT_DIR + filename + '.png')
+    plotScoutStackedTest(df_scout_fast_means_daiquiri_zkmixer,
+                suptitle="wasm compilers - optimizing (v8-turbofan) and single-pass (v8-liftoff) \n vs. \n wasm interpreter (wabt) with bignum host funcs",
+                suptitle_pos=1.07,
+                subtitle="daiquiri-zkmixer-websnark-bn128-groth16-four-pairings-and-mimc\n",
+                subtitle_size='xx-large')
 
-plotInterpOneEngine(df_interp, 'wasmi', all_tests, "Wasmi - all Precompiles (existing and proposed) compared")
-plotInterpOneEngine(df_interp, 'wabt', all_tests, "Wabt - all Precompiles (existing and proposed) compared")
+    plotInterpThreeTests(df_interp,
+                         blake2b_test_names,
+                         title="wasm interpreters compared - blake2b")
+
+    plotInterpThreeTests(df_interp,
+                         modexp_test_names,
+                         title="wasm interpreters compared - modexp")
+
+    plotInterpThreeTests(df_interp,
+                         bn128_add_test_names,
+                         title="wasm interpreters compared - bn128_add")
+
+    plotInterpThreeTests(df_interp,
+                         bn128_pairing_test_names,
+                         title="wasm interpreters compared - bn128_pairing")
+
+    plotThreeTestsExecTime(df_interp, blake2b_test_names, title="wasm interpreter execution time - blake2b")
+
+    plotThreeTestsExecTime(df_interp,
+                           modexp_test_names,
+                           title="wasm interpreter execution time - modexp")
+
+    plotThreeTestsExecTime(df_interp,
+                           bn128_mul_test_names,
+                           title="wasm interpreter execution time - bn128_mul")
+
+    plotThreeTestsExecTime(df_interp,
+                           bn128_pairing_test_names,
+                           title="wasm interpreter execution time - bn128_pairing")
+    plotCompilerSpeedup(df_wabt_v8liftoff, all_tests, interp_name="wabt", compiler_name="v8-liftoff")
+    plotCompilerSpeedup(df_fizzy_v8liftoff, all_tests, interp_name="fizzy", compiler_name="v8-liftoff")
+    plotCompilerSpeedup(df_wasm3_v8liftoff, all_tests, interp_name="wasm3", compiler_name="v8-liftoff")
+    plotCompilerSpeedup(df_wasmi_wavm, all_tests, interp_name="wasmi", compiler_name="wavm")
+
+    plotInterpOneEngine(df_interp, 'wasmi', all_tests, "Wasmi - all Precompiles (existing and proposed) compared")
+    plotInterpOneEngine(df_interp, 'wabt', all_tests, "Wabt - all Precompiles (existing and proposed) compared")
+
+
+    # plotCompilerStackedOneTest(df_compiler, "bls12-381-aggreg-128-pubkeys-verify-sig")
+    # plotCompilerStackedOneTest(df_native_and_compile, "bls12-381-aggreg-128-pubkeys-verify-sig", native=True)
+
+    df_scout_rolluprs_bn128_pairings = df_scout_data[df_scout_data['bench_name'] == 'ecpairing-zkrollup-rust-wasm-bn128-two-pairings']
+
+    df_rolluprs_native = df_scout_data[df_scout_data['bench_name'] == 'ecpairing-zkrollup-rust-native-bn128-two-pairings']
+    df_rolluprs_native = df_rolluprs_native.fillna(0)
+
+    df_scout_rolluprs = df_scout_rolluprs_bn128_pairings.append(df_rolluprs_native)
+
+    df_scout_rolluprs = df_scout_rolluprs[df_scout_rolluprs['engine'].isin(
+        ['rust-native', 'v8-liftoff', 'v8-turbofan', 'wabt-bignums-slowhost-slowmont', 'wabt-no-bignums']
+    )]
+        
+    df_scout_rolluprs.replace('wabt-bignums-slowhost-slowmont', 'wabt-with-bignums', inplace=True)
+    df_scout_means_rolluprs = df_scout_rolluprs.groupby(['engine']).mean()
+
+    output_file = 'rollup.rs-bn128-pairings-fast-scout-engines-v8-liftoff-and-wabt-with-bignums.png'
+    plotOneTestColoredTicks(df_scout_means_rolluprs,
+        output_file,
+        suptitle="rollup.rs-bn128-pairings - fast Scout engines (v8-liftoff and wabt-with-bignums)",
+        suptitle_pos=1.02,
+        subtitle="ecpairing-zkrollup-rust-wasm-bn128-two-pairings\n",
+        subtitle_size='xx-large',
+        highlight_tick="wabt-with-bignums")
+
+    df_scout_rolluprs_bn128_pairings = df_scout_data[df_scout_data['bench_name'] == 'ecpairing-zkrollup-rust-wasm-bn128-two-pairings']
+    df_rolluprs_native = df_scout_data[df_scout_data['bench_name'] == 'ecpairing-zkrollup-rust-native-bn128-two-pairings']
+    df_rolluprs_native = df_rolluprs_native.fillna(0)
+    df_scout_rolluprs = df_scout_rolluprs_bn128_pairings.append(df_rolluprs_native)
+
+    print(df_scout_rolluprs['engine'].unique().tolist())
+
+    df_scout_rolluprs = df_scout_rolluprs[df_scout_rolluprs['engine'].isin(
+        ['rust-native', 'v8-liftoff', 'v8-turbofan', 'wabt-bignums-slowhost-slowmont', 'wabt-bignums-slowhost-slowmont-superops']
+    )]
+
+    df_scout_rolluprs.replace('wabt-bignums-slowhost-slowmont', 'wabt-with-bignums', inplace=True)
+    df_scout_rolluprs.replace('wabt-bignums-slowhost-slowmont-superops', 'wabt-bignums-superops', inplace=True)
+
+    df_scout_means_rolluprs = df_scout_rolluprs.groupby(['engine']).mean()
+
+    plotOneTestColoredTicks(df_scout_means_rolluprs,
+        output_file = 'rust-wasm-bn128-pairing-super-ops-v8-liftoff-and-wabt.png',
+        suptitle="rollup.rs-bn128-pairings - fast Scout engines (v8-liftoff and wabt-with-bignums)",
+        suptitle_pos=1.02,
+        subtitle="ecpairing-zkrollup-rust-wasm-bn128-two-pairings\n",
+        subtitle_size='xx-large',
+        highlight_tick="wabt-bignums-superops")
+
+    df_scout_rust_vs_websnark = df_scout_data[
+        df_scout_data['bench_name'].isin(
+            ['ecpairing-zkrollup-rust-wasm-bn128-two-pairings',
+             'ecpairing-zkrollup-websnark-bn128-two-pairings']
+        )
+    ]
+
+    df_scout_rust_vs_websnark['engine'].unique().tolist()
+
+    df_scout_rust_vs_websnark = df_scout_data[
+        df_scout_data['bench_name'].isin(
+            ['ecpairing-zkrollup-rust-wasm-bn128-two-pairings',
+             'ecpairing-zkrollup-websnark-bn128-two-pairings']
+        )
+    ]
+
+# simplify bench_names to rust-wasm and websnark-wasm
+    df_scout_rust_vs_websnark.replace('ecpairing-zkrollup-rust-wasm-bn128-two-pairings', 'rust-wasm', inplace=True)
+    df_scout_rust_vs_websnark.replace('ecpairing-zkrollup-websnark-bn128-two-pairings', 'websnark-wasm', inplace=True)
+
+
+    df_scout_rust_vs_websnark = df_scout_rust_vs_websnark[df_scout_rust_vs_websnark['engine'].isin(
+        ['rust-native', 'v8-liftoff', 'v8-turbofan', 'wabt-bignums-slowhost-slowmont', 'wabt-bignums-slowhost-slowmont-superops']
+    )]
+
+# simplify wabt engine names
+    df_scout_rust_vs_websnark.replace('wabt-bignums-slowhost-slowmont', 'wabt-with-bignums', inplace=True)
+    df_scout_rust_vs_websnark.replace('wabt-bignums-slowhost-slowmont-superops', 'wabt-bignums-superops', inplace=True)
+
+    output_file = 'rust-vs-websnark-bn128-pairing.png'
+
+    plotOneTestGrouped(df_scout_rust_vs_websnark,
+        output_file,
+        title="websnark vs rust - zkrollup bn128 pairings (websnark/bn128 vs rollup.rs)",
+        test_title="",
+        group_order=['websnark-wasm', 'rust-wasm'],
+        sort_by=['rust-wasm'],
+        colors=['tab:orange', 'tab:blue'])
+
+    df_scout_websnark = df_scout_data[df_scout_data['bench_name'] == 'ecpairing-zkrollup-websnark-bn128-two-pairings']
+    df_rolluprs_native = df_scout_data[df_scout_data['bench_name'] == 'ecpairing-zkrollup-rust-native-bn128-two-pairings']
+    df_rolluprs_native = df_rolluprs_native.fillna(0)
+    df_scout_websnark = df_scout_websnark.append(df_rolluprs_native)
+
+    df_scout_websnark = df_scout_websnark[df_scout_websnark['engine'].isin(
+        ['rust-native', 'v8-liftoff', 'v8-turbofan', 'wabt-bignums-slowhost-slowmont',
+         'wabt-bignums-slowhost-slowmont-superops', 'wabt-bignums-slowhost-slowmont-superops',
+        'wabt-bignums-fasthost-slowmont-superops', 'wabt-bignums-fasthost-fastmont-superops']
+    )]
+
+    df_scout_websnark.replace('wabt-bignums-slowhost-slowmont', 'wabt-with-bignums', inplace=True)
+    df_scout_websnark.replace('wabt-bignums-slowhost-slowmont-superops', 'wabt-bignums-superops', inplace=True)
+    df_scout_websnark.replace('wabt-bignums-fasthost-slowmont-superops', 'wabt-bignums-superops-fasthost', inplace=True)
+    df_scout_websnark.replace('wabt-bignums-fasthost-fastmont-superops', 'wabt-bignums-superops-fasthost-fastmont', inplace=True)
+
+    output_file = 'websnark-bn128-pairing-fast-host-fast-mont.png'
+    df_scout_means_websnark = df_scout_websnark.groupby(['engine']).mean()
+    plotOneTestColoredTicks(df_scout_means_websnark,
+        output_file,
+        suptitle="websnark-bn128-pairings - engines compared (v8-liftoff and wabt-with-bignums)",
+        suptitle_pos=1.02,
+        subtitle="ecpairing-zkrollup-websnaark-bn128-two-pairings\n",
+        subtitle_size='xx-large',
+        highlight_tick="wabt-bignums-superops-fasthost-fastmont")
+
+    df_scout_websnark = df_scout_data[df_scout_data['bench_name'] == 'ecpairing-zkrollup-websnark-bn128-two-pairings']
+    df_rolluprs_native = df_scout_data[df_scout_data['bench_name'] == 'ecpairing-zkrollup-rust-native-bn128-two-pairings']
+    df_rolluprs_native = df_rolluprs_native.fillna(0)
+    df_scout_websnark = df_scout_websnark.append(df_rolluprs_native)
+
+    df_scout_websnark['engine'].unique().tolist()
+
+    df_scout_websnark = df_scout_data[df_scout_data['bench_name'] == 'ecpairing-zkrollup-websnark-bn128-two-pairings']
+    df_rolluprs_native = df_scout_data[df_scout_data['bench_name'] == 'ecpairing-zkrollup-rust-native-bn128-two-pairings']
+    df_rolluprs_native = df_rolluprs_native.fillna(0)
+    df_scout_websnark = df_scout_websnark.append(df_rolluprs_native)
+
+#df_scout_websnark.replace('ecpairing-zkrollup-rust-native-bn128-two-pairings', 'rust-native', inplace=True)
+#df_scout_websnark.replace('ecpairing-zkrollup-websnark-bn128-two-pairings', 'websnark-wasm', inplace=True)
+
+    df_scout_websnark = df_scout_websnark[df_scout_websnark['engine'].isin(
+        ['rust-native', 'v8-liftoff', 'v8-turbofan', 'wabt-bignums-slowhost-slowmont',
+         'wabt-bignums-slowhost-slowmont-superops', 'wabt-bignums-slowhost-slowmont-superops',
+        'wabt-bignums-fasthost-slowmont-superops']
+    )]
+
+    df_scout_websnark.replace('wabt-bignums-slowhost-slowmont', 'wabt-with-bignums', inplace=True)
+    df_scout_websnark.replace('wabt-bignums-slowhost-slowmont-superops', 'wabt-bignums-superops', inplace=True)
+    df_scout_websnark.replace('wabt-bignums-fasthost-slowmont-superops', 'wabt-bignums-superops-fasthost', inplace=True)
+
+    df_scout_means_websnark = df_scout_websnark.groupby(['engine']).mean()
+
+    output_file = 'websnark-bn128-pairing-fast-host.png'
+    plotOneTestColoredTicks(df_scout_means_websnark,
+        output_file,
+        suptitle="websnark-bn128-pairings - engines compared (v8-liftoff and wabt-with-bignums)",
+        suptitle_pos=1.02,
+        subtitle="ecpairing-zkrollup-websnaark-bn128-two-pairings\n",
+        subtitle_size='xx-large',
+        highlight_tick="wabt-bignums-superops-fasthost")
+
+
+plotCompilerStackedThreeTests(df_compiler, blake2b_test_names, title='Wasm Compilers Blake2b')
+plotCompilerStackedThreeTests(df_compiler, bn128_mul_test_names, title='Wasm Compilers bn128_mul')
+plotCompilerStackedThreeTests(df_compiler, bn128_pairing_test_names, title='Wasm Compilers bn128_pairing')
