+++ scripts/plot_util.py
@@ -1,0 +1,456 @@
+import matplotlib.pyplot as plt
+import matplotlib.ticker as ticker
+import pandas as pd
+import numpy as np
+import durationpy
+import math
+
+from os.path import join
+from adjustText import adjust_text
+
+IMG_OUTPUT_DIR = "../charts/"
+COLORS_DEFAULT = {'blue': '#348ABD', 'red': '#E24A33'}
+
+plt.style.use('ggplot')
+plt.rcParams.update({'figure.max_open_warning': 0})
+
+def toMs(secs):
+    if secs < 0.001:
+        μs = secs * 1000000
+        return "{}μs".format(round(μs, 1))
+    if secs < 0.1:
+        ms = secs * 1000
+        return "{}ms".format(round(ms, 1))
+    elif secs < 1:
+        ms = secs * 1000
+        return "{}ms".format(int(round(ms, 0)))
+    else:
+        return "{}s".format(round(secs, 2))
+
+def formatFileName(graph_title):
+    return '-'.join(graph_title.lower().split(' '))
+
+def plotInterpOneEngine(df_benchdata, engine, test_names, title=""):
+    filename = engine + '-all-precompiles-compared'
+    df_bench = df_benchdata.loc[engine][df_benchdata.loc[engine]['test_name'].isin(test_names)]
+    df_bench.set_index('test_name', inplace=True)
+    df_bench = df_bench.reindex(test_names)
+    plt.figure()
+    ax = df_bench[['parse_time', 'exec_time']].plot.bar(stacked=True, figsize=(15,7), color=[COLORS_DEFAULT['red'], COLORS_DEFAULT['blue']])
+    plt.ylabel("seconds")
+    ax.set(xlabel='')
+    # doing get_legend_handles_labels() so we can control the order of the labels in the legend 
+    handles, labels = ax.get_legend_handles_labels()
+    plt.legend([handles[1], handles[0]], ['execution time (top bar/label)', 'startup time (bottom bar/label - bar not visible for times < 20.0ms)'])
+    adjust_text_labels(labelBarHeights(ax))
+    plt.suptitle(title, fontsize=16, y=0.95)
+    print(f"plotInterpOneEngine: {filename}.png")
+    plt.savefig(IMG_OUTPUT_DIR + filename + '.png', bbox_inches='tight')
+
+def plotCompilerSpeedup(df_compiler_speedup, tests, interp_name="interp", compiler_name="compiler"):
+    filename = 'compiler-speedup-' + interp_name + '-' + compiler_name
+    df_is = df_compiler_speedup.copy()
+    plot_tests = df_is['test_name'].tolist()
+    test_index = tests.copy()
+    for test_name in tests:
+        if test_name not in plot_tests:
+            test_index.remove(test_name)
+ 
+    df_is = df_is.set_index('test_name')
+    df_is = df_is.reindex(test_index)
+    plt.figure()
+    ax = df_is['interp_slowdown'].plot.bar(figsize=(13,7), color=COLORS_DEFAULT['blue'])
+    ax.set_title("{} (Interpreter) vs {} (Compiler) - speedup".format(interp_name, compiler_name))
+    ax.set(ylabel="compiler speedup (log scale)", xlabel="")
+
+    y_max = ax.get_ylim()[1]
+    y_max = y_max * 5 # padding between top of bars and legend
+    ax.set_yscale('log')
+    ax.set_ylim(0.1, y_max) # adjustment to move horizontal line higher on chart
+
+    adjust_text_labels(labelBarHeights(ax, to_ms=False, to_factor=1, lower_y_bound=False))
+
+    plt.axhline(y=1.0, linewidth=2, ls='--', color='r')
+
+    handles, _ = ax.get_legend_handles_labels()
+
+    plt.legend(labels=["compiler engine is faster (ratio > 1x)", "compiler speedup \n (ratio: interp_exec_time/compiler_exec_time)"])
+
+    #f.suptitle(title, fontsize=16, y=0.95)
+    print(f"plotCompilerSpeedup: {filename}.png")
+    plt.savefig(IMG_OUTPUT_DIR + filename + '.png', bbox_inches='tight')
+
+# Interpreter engines compared
+def plotInterpThreeTests(df_testdata, three_names, title="Title", filter_engines=None):
+    filename = title.lower().replace(' ', '-')
+    if filter_engines is not None:
+        df_testdata = filterDfEngines(df_testdata.reset_index(), filter_engines)
+        df_testdata.set_index('engine', inplace=True)
+    df_1 = df_testdata[df_testdata['test_name'] == three_names[0]]
+    df_2 = df_testdata[df_testdata['test_name'] == three_names[1]]
+    df_3 = df_testdata[df_testdata['test_name'] == three_names[2]]
+    plt.figure()
+    f, (ax1) = plt.subplots(1, 3, sharey=True, figsize=(16,7))
+    df_1.sort_values('total_time')[['parse_time', 'exec_time']].plot.bar(ax=ax1[0], stacked=True)
+    ax1[0].set_title(three_names[0])
+    ax1[0].set(xlabel='')
+    df_2.sort_values('total_time')[['parse_time', 'exec_time']].plot.bar(ax=ax1[1], stacked=True)
+    ax1[1].set_title(three_names[1])
+    ax1[1].set(xlabel='')
+    df_3.sort_values('total_time')[['parse_time', 'exec_time']].plot.bar(ax=ax1[2], stacked=True)
+    ax1[2].set_title(three_names[2])
+    ax1[2].set(xlabel='')
+    adjust_text_labels(labelBarHeights(ax1[0]), ax=ax1[0])
+    adjust_text_labels(labelBarHeights(ax1[1]), ax=ax1[1])
+    adjust_text_labels(labelBarHeights(ax1[2]), ax=ax1[2])
+    ax1[0].legend(labels=["startup time", "execution time"])
+    ax1[1].legend(labels=["startup time", "execution time"])
+    ax1[2].legend(labels=["startup time", "execution time"])
+
+    ax1[1].tick_params(axis='y', left=True, labelleft=True)
+    ax1[2].tick_params(axis='y', left=True, labelleft=True)
+    
+    scale_y = 1e-3
+    ticks_y = ticker.FuncFormatter(lambda y, pos: '{0:g}'.format(y/scale_y))
+    ax1[1].yaxis.set_major_formatter(ticks_y)
+    
+    ax1[0].set(ylabel='milliseconds')
+    
+    f.suptitle(title, fontsize=16, y=0.98)
+    print(f"plotInterpThreeTests: {filename}.png")
+    plt.savefig(IMG_OUTPUT_DIR + filename + '.png', bbox_inches='tight')
+
+# Compiler function
+# Note: Fix scale for v8-turbofan execution time
+def plotCompilerStackedOneTest(df_benchdata, test_name, native=False):
+    filename = test_name
+    if native: filename += '-native'
+    df_1 = df_benchdata[df_benchdata['test_name'] == test_name]
+    # for some reason, exec_time needs to come first for the stacked chart to display correctly
+    ax = df_1[['exec_time', 'compile_time']].sort_values('exec_time').plot.bar(stacked=True, color=[COLORS_DEFAULT['blue'], COLORS_DEFAULT['red']], figsize=(10,10))
+    ax.set_title("Compilers - {}".format(test_name))
+    df_total_times = df_1[['exec_time', 'compile_time']].copy()
+    df_total_times['total_time'] = df_total_times['exec_time'] + df_total_times['compile_time']
+    ymax = max(df_total_times[['total_time']].max()) * 1.3 # 30% larger for padding between top of bars and top of graph
+    ymin = min([x for x in df_1[['exec_time', 'compile_time']].min().tolist() if x > 0])
+    ymin = ymin * 0.8 # 0.8 to get a number 20% smaller, fix to make bar appear for smallest exec time
+    ax.set_ylim(ymin, ymax)
+    ax.set_yscale("log")
+    plt.ylabel("seconds (log scale)")
+    adjust_text_labels(labelBarHeights(ax, lower_y_bound=False))
+    ax.legend(labels=["execution time", "compile time"])
+    print(f"plotCompilerStackedOneTest: {filename}.png")
+    plt.savefig(IMG_OUTPUT_DIR + 'wasm-compilers-' + filename + '.png', bbox_inches='tight')
+
+# Functions for plotting a two engine comparison
+def plotThreeTestsGrouped(df_benches, three_tests, title="Title"):
+    df_1 = df_benches[df_benches['bench_name'].str.contains(three_tests[0])]
+    df_2 = df_benches[df_benches['bench_name'].str.contains(three_tests[1])]
+    df_3 = df_benches[df_benches['bench_name'].str.contains(three_tests[2])]
+    # to group by engine, pivot to `columns='enginge'`
+    df_1 = pd.pivot_table(df_1, values='exec_time', columns=['engine'], index=['bench_name'], aggfunc=np.mean)
+    df_2 = pd.pivot_table(df_2, values='exec_time', columns=['engine'], index=['bench_name'], aggfunc=np.mean)
+    df_3 = pd.pivot_table(df_3, values='exec_time', columns=['engine'], index=['bench_name'], aggfunc=np.mean)
+    filename = title.lower().replace(' ', '-')
+
+    plt.figure()
+    f, (ax1) = plt.subplots(1, 3, sharey=True, figsize=(13,6))
+
+    df_1.sort_values('wabt-with-superops').plot.bar(ax=ax1[0])
+    ax1[0].set_title(three_tests[0])
+    ax1[0].set(ylabel='seconds', xlabel='')
+
+    df_2.sort_values('wabt-with-superops').plot.bar(ax=ax1[1])
+    ax1[1].set_title(three_tests[1])
+    ax1[1].set(xlabel='')
+
+    df_3.sort_values('wabt-with-superops').plot.bar(ax=ax1[2])
+    ax1[2].set_title(three_tests[2])
+    ax1[2].set(xlabel='')
+
+    adjust_text_labels(labelBarHeights(ax1[0]), ax=ax1[0])
+    adjust_text_labels(labelBarHeights(ax1[1]), ax=ax1[1])
+    adjust_text_labels(labelBarHeights(ax1[2]), ax=ax1[2])
+
+    f.suptitle(title, fontsize=16, y=0.98)
+    print(f"plotTreeTestsGrouped: {filename}.png ")
+    plt.savefig(IMG_OUTPUT_DIR + filename + '.png', bbox_inches='tight')
+
+def plotCompilerStackedThreeTests(df_benches, test_names, title='Title', native=False):
+    filename = title.lower().replace(' ', '-') # test_names[0]
+
+    if native:
+        filename += '-native'
+    plt.figure()
+
+    f, axs = plt.subplots(1, 3, sharey=True, figsize=(16, 7))
+
+    dfs = [
+        df_benches[df_benches['test_name'].str.contains(test_names[0])],
+        df_benches[df_benches['test_name'].str.contains(test_names[1])],
+        df_benches[df_benches['test_name'].str.contains(test_names[2])],
+    ]
+
+    for test_name, ax, df in zip(test_names, axs, dfs):
+        # for some reason, exec_time needs to come first for the stacked chart to display correctly
+        df[['exec_time', 'compile_time']].sort_values('exec_time').plot.bar(ax=ax, stacked=True, color=[COLORS_DEFAULT['blue'], COLORS_DEFAULT['red']])
+        ax.set_title("Compilers - {}".format(test_name))
+        # ax.set_title("Compilers - {}".format(test_name))
+        df_total_times = df[['exec_time', 'compile_time']].copy()
+        df_total_times['total_time'] = df_total_times['exec_time'] + df_total_times['compile_time']
+        ymax = max(df_total_times[['total_time']].max()) * 1.3 # 30% larger for padding between top of bars and top of graph
+
+        # ymin = min([x for x in df[['exec_time', 'compile_time']].min().tolist() if x > 0])
+        # ymin = ymin * 0.8 # 0.8 to get a number 20% smaller, fix to make bar appear for smallest exec time
+        # TODO don't hardcode this
+        ymin = 0.00001
+        ax.set_ylim(ymin, ymax)
+        ax.set_yscale("log")
+        plt.ylabel("seconds (log scale)")
+        adjust_text_labels(labelBarHeights(ax, lower_y_bound=False), ax=ax)
+        ax.legend(labels=["execution time", "compile time"])
+
+    print(f"plotCompilerStackedThreeTests: {filename}.png")
+    plt.savefig(IMG_OUTPUT_DIR + filename + '.png', bbox_inches='tight')
+
+# Interpreters - Execution time
+def plotThreeTestsExecTime(df_testdata, three_names, title="Title", filter_engines=None):
+    filename = title.lower().replace(' ', '-')
+    if filter_engines is not None:
+        df_testdata = filterDfEngines(df_testdata.reset_index(), filter_engines)
+        df_testdata.set_index('engine', inplace=True)
+    df_1 = df_testdata[df_testdata['test_name'] == three_names[0]]
+    df_2 = df_testdata[df_testdata['test_name'] == three_names[1]]
+    df_3 = df_testdata[df_testdata['test_name'] == three_names[2]]
+    plt.figure()
+    f, (ax1) = plt.subplots(1, 3, sharey=True, figsize=(16,7))
+    df_1.sort_values('exec_time')[['exec_time']].plot.bar(ax=ax1[0], color=COLORS_DEFAULT['blue'])
+    ax1[0].set_title(three_names[0])
+    ax1[0].set(xlabel='')
+    df_2.sort_values('exec_time')[['exec_time']].plot.bar(ax=ax1[1], color=COLORS_DEFAULT['blue'])
+    ax1[1].set_title(three_names[1])
+    ax1[1].set(xlabel='')
+    df_3.sort_values('exec_time')[['exec_time']].plot.bar(ax=ax1[2], color=COLORS_DEFAULT['blue'])
+    ax1[2].set_title(three_names[2])
+    ax1[2].set(xlabel='')
+    adjust_text_labels(labelBarHeights(ax1[0]), ax=ax1[0])
+    adjust_text_labels(labelBarHeights(ax1[1]), ax=ax1[1])
+    adjust_text_labels(labelBarHeights(ax1[2]), ax=ax1[2])
+    ax1[0].legend(labels=["execution time"])
+    ax1[1].legend(labels=["execution time"])
+    ax1[2].legend(labels=["execution time"])
+
+    ax1[1].tick_params(axis='y', left=True, labelleft=True)
+    ax1[2].tick_params(axis='y', left=True, labelleft=True)
+   
+    scale_y = 1e-3
+    ticks_y = ticker.FuncFormatter(lambda y, pos: '{0:g}'.format(y/scale_y))
+    ax1[1].yaxis.set_major_formatter(ticks_y)
+    
+    ax1[0].set(ylabel='milliseconds')
+    
+    f.suptitle(title, fontsize=16, y=0.98)
+    print(f"plotThreeTestsExecTime: {filename}.png")
+    plt.savefig(IMG_OUTPUT_DIR + filename + '.png', bbox_inches="tight")
+
+# Functions for plotting biturbo and bignum benchmarks
+# non-stacked bar chart, for only plotting exec time or total time
+def plotOneTest(df_data, suptitle="title", suptitle_pos=1.00, subtitle="subchart", subtitle_size='medium'):
+    filename = subtitle.strip().lower().replace(' ', '-')
+    plt.figure()
+    df_total_times = df_data[['exec_time', 'parse_time']].copy()
+    df_total_times.fillna(0, inplace=True)
+    df_total_times['time'] = df_total_times['exec_time'] + df_total_times['parse_time']
+    ax = df_total_times.sort_values('time')[['time']].plot.bar(figsize=(14,8), color=COLORS_DEFAULT['blue'])
+    # fontsize = {size in points, 'xx-small', 'x-small', 'small', 'medium', 'large', 'x-large', 'xx-large'}
+    ax.set_title(subtitle, fontsize=subtitle_size)
+    ax.set(ylabel="seconds", xlabel="")
+    ax.xaxis.set_tick_params(labelrotation=60, labelsize=12)
+    adjust_text_labels(labelBarHeights(ax))
+    plt.suptitle(suptitle, fontsize=18, y=suptitle_pos)
+    print(f"plotOneTest: {filename}.png")
+    plt.savefig(IMG_OUTPUT_DIR + filename + '.png', bbox_inches='tight')
+
+def plotScoutStackedTest(df_data, suptitle="title", suptitle_pos=1.00, subtitle="subchart", subtitle_size='medium'):
+    filename = subtitle.strip().lower().replace(' ', '-')
+    plt.figure()
+    # for some reason, exec_time needs to come first for the stacked chart to display correctly
+    ax = df_data[['exec_time', 'parse_time']].sort_values('exec_time').plot.bar(stacked=True, color=[COLORS_DEFAULT['blue'], COLORS_DEFAULT['red']], figsize=(10,10))
+    # fontsize = {size in points, 'xx-small', 'x-small', 'small', 'medium', 'large', 'x-large', 'xx-large'}
+    ax.set_title(subtitle, fontsize=subtitle_size)
+    ax.set(ylabel="seconds", xlabel="")
+    df_total_times = df_data[['exec_time', 'parse_time']].copy()
+    df_total_times['total_time'] = df_total_times['exec_time'] + df_total_times['parse_time']
+    ax.xaxis.set_tick_params(labelrotation=60, labelsize=12)
+    adjust_text_labels(labelBarHeights(ax, lower_y_bound=False))
+    ax.legend(labels=["execution time", "startup time"])
+    plt.suptitle(suptitle, fontsize=18, y=suptitle_pos)
+    print(f"plotScoutStackedTest: {filename}.png")
+    plt.savefig(IMG_OUTPUT_DIR + filename + '.png', bbox_inches='tight')
+
+def labelBarWidths(ax, to_ms=True, to_factor=False, to_kb=False, round_places=None):
+    for p in ax.patches:
+        label_val = p.get_width()
+        if label_val == 0:
+            continue
+        x_pos = p.get_x() + label_val
+        x_offset = -30
+        x_bound = ax.get_xbound()[1]
+        if (x_pos > x_bound):
+            x_pos = x_bound
+        if label_val < (0.15 * x_bound):
+            # if bar is too small to fit text, place next to bar
+            x_offset = 30
+        if label_val > 1000:
+            label_val = int(label_val)
+        if to_ms:
+            label_val = toMs(label_val)
+        if to_factor:
+            label_val = str(round(label_val, 2)) + "x"
+        if to_kb:
+            kbs = label_val / 1000.0
+            label_val = str(round(kbs, 1)) + " kB"
+        if round_places:
+            label_val = str(round(label_val, round_places))
+        ax.annotate(label_val,
+                    (x_pos,
+                     p.get_y()+p.get_height()/2.),
+                    ha='center',
+                    va='center',
+                    xytext=(x_offset, 0),
+                    textcoords='offset points',
+                    fontsize=10,
+                    weight="bold")
+
+def labelBarHeights(ax, to_ms=True, small=False, to_factor=False, lower_y_bound=True):
+    labels = []
+    for p in ax.patches:
+        if p.get_height() == 0:
+            continue
+        label_val = p.get_height()
+        y_pos = p.get_y() + label_val
+        y_offset = 0
+        y_bound = ax.get_ybound()[1]
+        if y_pos > y_bound:
+            y_pos = y_bound
+        label_val = p.get_height()
+        y_pos = p.get_y() + label_val
+        y_offset = 0
+        y_bound = ax.get_ybound()[1]
+        if y_pos > y_bound:
+            y_pos = y_bound
+        if lower_y_bound:
+            if label_val < (0.03 * y_bound) and small is False:
+                # don't place labels near bottom axis, adjust_text cant deal
+                y_offset = 0.03 * y_bound
+        
+        if label_val > 1000:
+            label_val = int(label_val)
+        if to_ms:
+            label_val = toMs(label_val)
+        if to_factor:
+            if not isinstance(to_factor, int):
+                to_factor = 2
+            label_val = str(round(label_val, to_factor)) + "x"
+        """
+        txt_label = ax.annotate(toMs(label_val),
+                    (p.get_x()+p.get_width()/2.,
+                     y_pos),
+                    ha='center',
+                    va='center',
+                    xytext=(0, y_offset),
+                    textcoords='offset points',
+                    fontsize=10,
+                    weight="bold")
+        """
+ 
+        txt_label = ax.text(
+                    x=p.get_x()+p.get_width()/2.,
+                    y=y_pos+y_offset,
+                    s=label_val,
+                    ha='center',
+                    va='center',
+                    #xytext=(0, y_offset),
+                    #textcoords='offset points',
+                    fontsize=10,
+                    weight="bold")
+
+        labels.append(txt_label)
+
+    return labels    
+
+def plotOneTestColoredTicks(df_data, output_file, suptitle="title", suptitle_pos=1.00, subtitle="subchart", subtitle_size='medium', highlight_tick=None):
+    plt.figure()
+    df_total_times = df_data[['exec_time', 'parse_time']].copy()
+    df_total_times.fillna(0, inplace=True)
+    df_total_times['time'] = df_total_times['exec_time'] + df_total_times['parse_time']
+    ax = df_total_times.sort_values('time')[['time']].plot.bar(figsize=(14,8), color=COLORS_DEFAULT['blue'])
+    # fontsize = {size in points, 'xx-small', 'x-small', 'small', 'medium', 'large', 'x-large', 'xx-large'}
+    ax.set_title(subtitle, fontsize=subtitle_size)
+    ax.set(ylabel="seconds", xlabel="")
+    ax.xaxis.set_tick_params(labelrotation=60, labelsize=12)
+    if highlight_tick is not None:
+        for i in plt.gca().get_xticklabels():
+            if i.get_text() == highlight_tick:
+                i.set_color('tab:green')
+                i.set_fontstyle('oblique')
+                i.set_fontweight('heavy')
+                
+        # {'tab:blue', 'tab:orange', 'tab:green', 'tab:red', 'tab:purple', 'tab:brown', 'tab:pink', 'tab:gray', 'tab:olive', 'tab:cyan'}
+
+    adjust_text_labels(labelBarHeights(ax))
+    plt.suptitle(suptitle, fontsize=18, y=suptitle_pos)
+    
+    plt.savefig(IMG_OUTPUT_DIR + output_file, bbox_inches='tight')
+
+def adjust_text_labels(labels, ax=None):
+    adjust_text(labels, ax=ax, autoalign='y', only_move={'points':'y', 'text':'y'})
+
+def read_results(file_name):
+    csv_results = pd.read_csv(file_name)    
+    return csv_results
+                       
+def filterDfEngines(df_data, engines):
+    df_data_filtered = df_data.copy()
+    for fil in engines:
+        df_data_filtered = df_data_filtered[df_data_filtered['engine'] != fil]
+
+    return df_data_filtered
+
+def add_engine_ratio_col(df_two_engines, interp_engine, compiler_engine):
+    df_two = df_two_engines.copy(deep=True)
+    df_interp = df_two[df_two['engine'] == interp_engine].copy()
+    df_compile = df_two[df_two['engine'] == compiler_engine]
+    for index, row in df_interp.iterrows():
+        test_name = row['test_name']
+        row_compiler = df_compile[df_compile['test_name'] == test_name]
+        if len(row_compiler) != 1:
+            print("ERROR. SHOULD HAVE 1 MATCH ONLY")
+            break
+        compiler_exec_time = row_compiler.iloc[0]['exec_time']
+        exec_ratio = row['exec_time'] / compiler_exec_time
+        df_interp.at[index, 'compiler_exec_time'] = compiler_exec_time
+        df_interp.at[index, 'interp_slowdown'] = exec_ratio
+
+    return df_interp
+
+def plotOneTestGrouped(df_1, output_file, test_title="test_title", title="plot Title", colors=['tab:blue', 'tab:orange'], sort_by='websnark-bn128-two-pairings', group_order=None):
+    # to group by engine, pivot to `columns='engine'`
+    #df_1 = pd.pivot_table(df_1, values='exec_time', columns=['engine'], index=['bench_name'], aggfunc=np.mean)
+    # group by bench name
+    df_1['time'] = df_1['exec_time'] + df_1['parse_time']
+    df_1 = pd.pivot_table(df_1, values='time', columns=['bench_name'], index=['engine'], aggfunc=np.mean)
+    # order to get websnark as the group on the left
+    if group_order is not None:
+        df_1 = df_1[group_order]
+
+    plt.figure()
+    ax = df_1.sort_values(sort_by).plot.bar(figsize=(14,8), color=colors)
+    #ax = df_1.plot.bar(figsize=(14,8))
+    ax.set_title(test_title)
+    ax.set(ylabel='seconds', xlabel='')
+
+    adjust_text_labels(labelBarHeights(ax))
+
+    plt.suptitle(title, fontsize=16, y=0.98)
+    plt.savefig(IMG_OUTPUT_DIR + output_file, bbox_inches='tight')
